mental-health-backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
motor==3.3.2
pymongo==4.6.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4

# AI and ML dependencies
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
tokenizers==0.15.0

# Text processing and emoji support
emoji==2.8.0
regex==2023.10.3

# PDF Generation
reportlab==4.0.7
matplotlib==3.8.2
pillow==10.1.0

# HTTP and CORS
httpx==0.25.2
python-multipart==0.0.6

# Data validation
pydantic[email]==2.5.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1